Add files via upload
This file contains the functions to display the training history, plot the ROC Curves and to compute the rejection ratios. It is unfortunately necessary to import the outputs of the model from the file "JetClass_JetTagging" since we were unable to resolve the numpy dependency issues within the same file. Therefore, we train the model in the "JetClass_JetTagging" file and then use the outputs of the model in the "JetClass_evaluationMetrics" file. The outputs of our model are stored in the files "training_history.json", "y_probs.json" and "y_true.json".